su, http, nth, sip, tport: removed some warnings.
darcs-hash:20061128195213-65a35-210b321f01bc3820a94b62d91fb2d378307042e5.gz

+++ libsofia-sip-ua/http/test_http.c
@@ -1197,7 +1197,7 @@ static int test_http_encoding(void)
   http_t *http;
   su_home_t *home;
   char b[160];
-  int n;
+  size_t n;
 
   BEGIN();
